--- a/go.mod
+++ b/go.mod
@@ -5,16 +5,17 @@ go 1.22.11
 toolchain go1.23.5
 
 require (
-	github.com/cespare/xxhash/v2 v2.3.0
 	github.com/dgryski/go-farm v0.0.0-20240924180020-3414d57e47da
 	github.com/dustin/go-humanize v1.0.1
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.10.0
+	github.com/zeebo/xxh3 v1.0.3-0.20230502181907-7a4a65f1295e
 	golang.org/x/sys v0.30.0
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/klauspost/cpuid/v2 v2.0.9 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )